txscript: Optimize typeOfScript pay-to-pubkey-hash.
This continues the process of converting the typeOfScript function to
use a combination of raw script analysis and the new tokenizer instead
of the far less efficient parsed opcodes.

In particular, it converts the detection of pay-to-pubkey-hash scripts
to use raw script analysis.

In order to accomplish this, it introduces two new functions.  The first
one is named extractPubKeyHash and works with the raw script bytes
to simultaneously determine if the script is a pay-to-pubkey-hash script,
and in the case it is, extract and return the hash.  The second new
function is named isPubKeyHashScript and is defined in terms of the
former.

The extract function approach was chosen because it is common for
callers to want to only extract relevant details from a script if the
script is of the specific type.  Extracting those details requires
performing the exact same checks to ensure the script is of the correct
type, so it is more efficient to combine the two into one and define the
type determination in terms of the result so long as the extraction does
not require allocations.

The following is a before and after comparison of analyzing a large
script:

benchmark                     old ns/op    new ns/op    delta
----------------------------------------------------------------
BenchmarkIsPubKeyHashScript   165903       0.64         -100.00%

benchmark                     old allocs   new allocs   delta
----------------------------------------------------------------
BenchmarkIsPubKeyHashScript   1            0            -100.00%

benchmark                     old bytes    new bytes    delta
----------------------------------------------------------------
BenchmarkIsPubKeyHashScript   466945       0            -100.00%